requirements.txt
streamlit
Keras
tensorflow==2.8.1
protobuf==3.20.0
Pillow
numpy
